feat: extract tests into separate modules
