+++ jsgenerator-test/jsgenerator-test-ante/pom.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <artifactId>jsgenerator-test</artifactId>
+        <groupId>com.osscameroon</groupId>
+        <version>${revision}</version>
+    </parent>
+    <modelVersion>4.0.0</modelVersion>
+
+    <artifactId>jsgenerator-test-ante</artifactId>
+
+    <dependencies>
+        <dependency>
+            <groupId>com.osscameroon</groupId>
+            <artifactId>jsgenerator-ante</artifactId>
+        </dependency>
+    </dependencies>
+</project>
