+++ jsgenerator-test/jsgenerator-test-core/src/test/resources/htmlFilesInput/sample.html
